+++ laktory/models/__init__.py
@@ -1,6 +1,8 @@
 from .basemodel import BaseModel
 from .dataevent import DataEvent
+from .dataframeexpression import DataFrameExpression
 from .dataproducer import DataProducer
+from .dataquality import *
 from .datasinks import *
 from .datasources import *
 from .grants import *
@@ -8,5 +10,5 @@ from .pipeline import Pipeline
 from .pipelinenode import PipelineNode
 from .pipelinenodeexpectation import PipelineNodeExpectation
 from .resources import *
-from .transformers import *
 from .stacks import *
+from .transformers import *